+++ genomenotebook/track.py
@@ -3,7 +3,7 @@
 # %% auto 0
 __all__ = ['Track']
 
-# %% ../nbs/API/01_track.ipynb 4
+# %% ../nbs/API/01_track.ipynb 5
 from fastcore.basics import *
 
 from bokeh.plotting import figure
@@ -28,7 +28,7 @@ import warnings
 
 
 
-# %% ../nbs/API/01_track.ipynb 7
+# %% ../nbs/API/01_track.ipynb 8
 class Track:
     """ Track objects should only be created through GenomeBrowser.add_track """
     def __init__(self,
@@ -49,7 +49,7 @@ class Track:
         
 
 
-# %% ../nbs/API/01_track.ipynb 12
+# %% ../nbs/API/01_track.ipynb 13
 @patch
 def _set_track_data_source(self:Track, data, pos, columns):
     columns=[c for c in columns if c] #some arguments can be None => remove them
@@ -80,7 +80,7 @@ def _set_track_data_source(self:Track, data, pos, columns):
     self.fig.x_range.js_on_change('start', xcb)
 
 
-# %% ../nbs/API/01_track.ipynb 13
+# %% ../nbs/API/01_track.ipynb 14
 @patch
 def line(self:Track,
          data: pd.DataFrame, #pandas DataFrame containing the data
@@ -92,10 +92,10 @@ def line(self:Track,
     self.fig.line(source=self.loaded_data, x=pos, y=y, **kwargs)
 
 
-# %% ../nbs/API/01_track.ipynb 16
+# %% ../nbs/API/01_track.ipynb 17
 from bokeh.transform import factor_cmap
 
-# %% ../nbs/API/01_track.ipynb 17
+# %% ../nbs/API/01_track.ipynb 18
 @patch
 def scatter(self:Track,
          data: pd.DataFrame, #pandas DataFrame containing the data
@@ -117,7 +117,7 @@ def scatter(self:Track,
         self.fig.scatter(source=self.loaded_data, x=pos, y=y, **kwargs)
 
 
-# %% ../nbs/API/01_track.ipynb 22
+# %% ../nbs/API/01_track.ipynb 23
 @patch
 def bar(self:Track,
          data: pd.DataFrame, #pandas DataFrame containing the data